Use URandom for Random DNS Query ID - P1
Instead of using a simple counter, try use /dev/urandom if available
to get generate the ID which will be sent as part of dns query.

The idea here is to make the life of a attacker who may be trying
to pollute the DNS resolution info maintained by the dns resolver
of uClibc, difficult by not allowing them to easily guess the ID
they need to use in the DNS response sent as part of a DNS poison
attack.

This is in response to ICS-VU-638779, CERT/CC VU#473698 message
seen on the devel mailing list.