Resolv: Trap post increament bug possibility, change to pre incr
Purposefully changed to /dev/urandom1 to verify the possible issue
with post increment of local_id passed to dnsrand_next. As had
thought, the local_id wont increment in this case, bcas urandom
fails and fallsback to passed default_value which is the local_id
before increment, so dns query id will be stuck at 1 or what ever
the initially set value.

My C memory has become bit rustic ;), even thou the doubt was there
in the background.

This is fixed by changing to pre increment before passing as defVal
to dnsrand_next.