Resolv: Let DnsRandNext be (pseudo) random most of the time
Previously if /dev/urandom was not accessible for what ever reason
it would have fallen back to a simple counter. Now instead it will
try to use realtime clock if available, to help with the periodic
reseeding. But eitherway it will chug along with prng one way or the
other, with reseeding or otherwise.