Resolv: xor short run, incase int64_t is not supported on target
If a given target doesnt support int64_t type, then the large mult
with mod operation cant be used to generate opaque random numbers
from the internal random sequence. In this case option for a xor
based random number transform logic is provided, but in prng short
run mode, which needs to be enabled. This should be ok in general.

In case of the opaque / one way transform based prng run, increase
interval between reseeding.