Resolv+: Message/Note update wrt xor, also clock alert in c also
As the transform on internal random data before exposing to outside
may or may not be one way (based on whether int64_t is supported
or not), The same is mentioned as part of Config help.

Also the possible corner cases with use of clock strategy, as io
and or processor performance (speed/multi/...) improves is alerted
in code comment also now.